requirements.txt
SpeechRecognition>=3.8.1
pyttsx3>=2.90
openai>=1.0.0
python-dotenv>=1.0.0
PyAudio>=0.2.11
rich>=13.0.0